added alibaba and puzzles solution

+++ Module 6.5/D_Ali_Baba_and_Puzzles.c
@@ -2,10 +2,10 @@
 
 #include<stdio.h>
 int main(){
-    int a,b,c;
+    long long int a,b,c;
     long long int d;
-    scanf("%d %d %d %lld",&a, &b, &c, &d);
-    if (a + b - c == d || a - b + c == d ||  a + b * c == d || a * b + c  == d || a - b * c == d || a * b - c == d)
+    scanf("%lld %lld %lld %lld",&a, &b, &c, &d);
+    if (a + b - c == d || a - b + c == d ||  a + (b * c) == d || (a * b) + c  == d || a - (b * c) == d || (a * b) - c == d)
     {
         printf("YES");
     }else{
